requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.27.0
anyio==3.7.1
httpcore==1.0.2
urllib3==2.0.7
python-dotenv==1.0.0
